Fix Error: Failed to load config "maven" to extend from.
Fix error when extending "maven/eslint-ts" when developing eslint-config-maven (depending on it using yarn link, npm link, or "eslint-config-maven": "file:../eslint-config-maven").

The cause of the error is that eslint-ts.js says "extends": ["maven"], it results in require('module').createRequire('/path/to/eslint-config-maven/eslint-ts.js').resolve('eslint-config-maven'). This works to locate its own directory as long as the path contains node_modules (since nodejs looks for subdirectories of node_modules). But when the parent of eslint-config-maven is not node-modules, then a module cannot re-import its own library using a require call that does not contain ./.