Publish the package to GitHub packages under @themaven-net
Change the instructions to use the GitHub npm registry under the @themaven-net organization instead of to npm install the git url. Since other packages are likely to use the GitHub package registry (e.g. to load js packages that were compiled from TypeScript), it is probably more uniform to add a dependency on the package instead of the source.